Creating vito.xml from address files.
We can now filter a vclient log file with filter_vclient_log.py
and have it write out the readable memory addresses
with the "-a" option. The resulting file can be read by
create_vito.py now, such that a vito.xml with only the
accessible memory addresses is created.